# openapi.yaml
openapi: 3.0.3
info:
  title: BankWiz API
  version: 0.1.0
  description: An API for accessing banking functionalities provided by BankWiz
servers:
  - url: http://localhost:8080
    description: Local development (server)
...
security:
  - oauth2:
    - openid
    - profile
    - email
paths:
...
  /group:
    post:
      tags:
        - GroupService
      summary: Create a group
      operationId: createGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupCreationRequest'
      responses:
        '201':
          description: Group created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupIndexDTO'
        '400':
          description: Invalid request. Please check the provided data.
        default:
          description: Unexpected error
  /group/{id}:
    get:
      tags:
        - GroupService
      summary: Get a group details
      operationId: getGroupDetails
      parameters:
        - name: id
          in: path
          description: Group ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Get group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupDetailsDTO'
        '400':
          description: Invalid request. Please check the provided data.
        default:
          description: Unexpected error
    delete:
      tags:
        - GroupService
      summary: Delete a group
      operationId: deleteGroup
      parameters:
        - name: id
          in: path
          description: Group ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Group deleted
        '400':
          description: Invalid request. Please check the provided data.
        default:
          description: Unexpected error
  /group/{groupId}/user:
    post:
      tags:
        - GroupService
      summary: Add user to group
      operationId: addUserGroup
      parameters:
        - name: groupId
          in: path
          description: Group ID
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddUserGroupRequest'
      responses:
        '200':
          description: User added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserGroupRightDTO'
        '400':
          description: Invalid request. Please check the provided data.
        default:
          description: Unexpected error
  /group/{groupId}/user/{userId}:
    delete:
      tags:
        - GroupService
      summary: Delete user from a group
      operationId: deleteUserFromGroup
      parameters:
        - name: groupId
          in: path
          description: Group ID
          required: true
          schema:
            type: string
            format: uuid
        - name: userId
          in: path
          description: User ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User deleted
        '400':
          description: Invalid request. Please check the provided data.
        default:
          description: Unexpected error
  /group/groups:
    get:
      tags:
        - GroupService
      summary: Get all groups of user
      operationId: getUserGroups
      responses:
        '200':
          description: Get all groups of user
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: '#/components/schemas/GroupIndexDTO'
        '400':
          description: Invalid request. Please check the provided data.
        default:
          description: Unexpected error
#################################### GroupService END ############################### 
#################################### BankAccountService START ####################### 
...
components:
  securitySchemes:
    oauth2:   # <--- Arbitrary name for the security scheme. Used to refer to it from elsewhere.
      type: oauth2
      description: For more information, see https://api.slack.com/docs/oauth
      flows: 
        implicit:
          authorizationUrl: https://bankwiz-dev.eu.auth0.com/authorize?audience=Bankwiz_server
          scopes:
            openid: openid
            profile: profile
            email: email
  schemas:
...
    GroupCreationRequest:
      type: object
      properties:
        groupName:
          type: string
      required:
        - groupName
...
    UserDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
          example: johndoe@example.com
      required:
        - id
        - email
    AddUserGroupRequest:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        right:
          $ref: '#/components/schemas/UserGroupRightEnum'
      required:
        - userId
        - right
...
    GroupIndexDTO:
      type: object
      properties:
        groupId:
          type: string
          format: uuid
        groupName:
          type: string
      required:
        - groupId
        - groupName
    UserGroupRightDTO:
      type: object
      properties:
        groupRightId:
          type: string
          format: uuid
        user:
          $ref: '#/components/schemas/UserDTO'
        right:
          $ref: '#/components/schemas/UserGroupRightEnum'
      required:
        - groupRightId
        - user
        - right
    GroupDetailsDTO:
      type: object
      properties:
        groupId:
          type: string
          format: uuid
        groupName:
          type: string
        usersRights:
          type: array
          items:
            $ref: '#/components/schemas/UserGroupRightDTO'
      required:
        - groupId
        - groupName
        - usersRights
...
    UserGroupRightEnum:
      type: string
      enum:
        - READ
        - WRITE
        - ADMIN
...